pyproject.toml
[project]
name = "boulder-log"
version = "0.1.0"
description = "Django app for logging boulder problems."
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "django>=4.2.19",
    "matplotlib>=3.10.1",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "pillow-heif>=0.22.0",
    "ruff>=0.9.6",
]

[tool.mypy]
files = "."
exclude = ["build", "migrations"]
plugins = ["mypy_django_plugin.main"]

# Use strict defaults
strict = true
warn_unreachable = true
warn_no_return = true

[[tool.mypy.overrides]]
# Don't require test functions to include types
module = "tests.*"
allow_untyped_defs = true
disable_error_code = "attr-defined"

[[tool.mypy.overrides]]
# Ignore packages with no stubs.
module = [
    'json_log_formatter',
]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "boulder_log.core.settings"

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "jupyterlab>=4.3.6",
]
